requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
requests